requirements.txt
kivy==2.2.1
pysocks==1.7.1
pyaes==1.6.1
ecdsa==0.18.0
requests==2.31.0
pillow==10.1.0
